pyproject.toml
[project]
name = "shortcomings-cli"
version = "0.0.1"
description = "Maintain and search project features and shortcomings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.24.0",
    "pyyaml>=6.0",
]

[project.scripts]
shortcomings = "shortcomings.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "ty>=0.0.19",
]